[NFC] CMPLRLLVM-31762: Replace use of "erase" in MapVector with "remove_if". (#670)
In the previous change set, I replaced a map with a MapVector. This introduced
a bug because, as Erich Keene explained, a class with a vector that is iterated
over has its iterators invalidated when erase() is called using an iterator.
To avoid this issue, we can, as Chris explained, use the remove_if() method.

I have replaced the use of erase() with MapVector in this file with remove_if().
There is still one use of erase() in the file, but that should not be a problem,
as the iterators are not used after erase() is called.

In remapCBVec, I removed the erase() withgout using remove_if() because the
erase() is not needed because assigning CBNewMap to CBVec[Index] causes the
destructor to be called on the old CBMap, making an explicit call to erase()
unneeded.

I am marking this as NFC as the issue is a random one using invalid iterators
and is not reproducible reliably.