[DTrans] Fix regression/assert with bitfield as last member (#732)
The previous DTrans patch attempted to detect padding after a field, but
ended up asserting if the bitfield was the last llvm value when attempting to
find tail padding.  This patch corrects that behavior by only checking
if we know there is a potential padding field after it.

Additionally, during diagnosing this, I found an empty array assert,
since empty-global-arrays don't have any initializers. However, these
aren't decomposed the same way that ones with initializers are, so we
can just use the type from the clang array for these.

This fixes CMPLRLLVM-31911