[SOAToAOSOP] (CMPLRLLVM-30335): Changes to port classification and transformation of Array methods for SOAToAOSOP. (#653)
There is no major change in the original implementation of SOAToAOS except
1. Make necessary changes to work opaque pointers
2. With opaque pointers, cloning is not occurred in DTransOPOptBase
for Array Methods because function types are same with and without
transformations due to typeless “ptr”. To reuse the existing code,
a temporary clone is created (in postprocessFunction).
Apply transformations for IR in the cloned routine and move the IR to the
OrigFunc. Ans, then delete the temporary cloned routine.
3. In addition to StoreInst, LoadInst also cloned for all elements before
mutating types to avoid issues with opaque pointers.