review/maksimsab/private/maksimsam/cherry pick (#824)
* [Loopopt][CMPLRLLVM-31927] Fix loopopt lit tests in xmain-web due to sext printing

* CMPLRLLVM-31902: Update an autogen SCEV test for Intel improvements (#763)

flags-from-poison.ll originally had hand-written checks in it, but
upstream converted it to an autogen test. When this was merged in, the
checks became more specific and started failing due to additional
no-wrap flags proved by xmain.

This change uses autogen to update the test for xmain.

* intel-noalias-args.ll: Fix bad alias.scope format. vecreduce.ll: icx can constant fold add-reductions. (#663)

* intel-noalias-args.ll: Fix bad alias.scope format.
vecreduce.ll: icx can constant fold add-reductions.

* Add missing CHECK-NEXT.

* CMPLRLLVM-31900: Remove test case intel-opt-LTO-pipeline-sw-advanced.ll (#693)

The community started to remove flags and test cases related to
the legacy pass manager. The test case opt-LTO-pipeline.ll was
removed in this change set: https://reviews.llvm.org/D109180 .
It used some flags that aren't supported anymore
(https://reviews.llvm.org/D109664).

This change set removed the test case
intel-opt-LTO-pipeline-sw-advanced.ll since it was the same test
as opt-LTO-pipeline.ll, but it only checks for the passes
that are guarded with INTEL_FEATURE_SW_ADVANCED.

Co-authored-by: Chenyang Liu <chenyang.liu@intel.com>
Co-authored-by: tcreech-intel <timothy.m.creech@intel.com>
Co-authored-by: Chang Lin <chang-sun.lin.jr@intel.com>
Co-authored-by: Axel Y. Rivera <axel.y.rivera@intel.com>