CMPLRLLVM-30334: Merge types using opaque pointers and DTrans metadata (Part 2) (#779)
* CMPLRLLVM-30334: Merge types using opaque pointers and DTrans metadata (Part 2)

This change set introduces the use of DTrans metadata for
merging structures with opaque pointers. The basic idea is
to check when a field in the source module and in the destination
module is a pointer, then we query the DTrans metadata to collect
the element type and map them.

Also, this change set fixes an issue with anonymous structures
where two anonymous structures with the same naming could be
mapped but in reality they shouldn't.

* This fix adds the check when the fields are arrays/vectors of
pointers, extra check if structures are packed, more test cases
and grammar errors.