CMPLRLLVM-20091: Fix for segfault when analyzing global dope vectors [XMAIN-REL] (#789)
This change set fixes a segfault that occurs if the information
for a global dope vector is collected except the alloc site. In
that case, it will trigger the analysis process and could hit
a segfault.

This change set guards the analysis process to check first for
alloc sites.

These are the same changes as

https://github.com/intel-restricted/applications.compilers.llvm-project/pull/707

but for xmain-rel.