Do not call constructor/destructor for firstprivate structs on target. (#798)
Since WriteArgumentsToHomeLocations code was removed from xmain,
we can now avoid calling copy constructor/destructor for firstprivate
non-pod objects on target constructs, if we are optimizing it into
a kernel argument.

The problem with calling constructor/destructor for Kokkos/HotQCD
is that this causes pulling in C++ STD methods that are not available
on SPIR-V devices. It is still a problem with such firstprivate objects
for other OpenMP constructs (e.g. teams), but this may be worked around
on the source level. For example,

Instead of:
One can use:

Then non_pod_obj will be shared for "teams ...", and constructor/destructor
will not be called inside the whole target region.

This change-set resolves CMPLRLLVM-32030 (given that additional source
changes in the test are made).