CMPLRLLVM-32051: Temporary fix for xmain-cand to enable "-opaque-pointers" in llvm-link (#906)
The community recently made mandatory to use the flag
"-opaque-pointers" when pointer types are disabled
(https://reviews.llvm.org/D109290). The tool llvm-link
creates its own llvm::Context, but it doesn't enable
the opaque pointers support, neither the flag works.
This fix that will try to check if "-opaque-pointers"
was specified in the command line and will enable the
use for opaque pointers. This is a temporary solution
until the community decides to address this issue.