[DTrans - OpaquePtr] DTransSafetyAnalyzer to analyze initializers with GEPOperators (#897)
The safety analyzer checks for global variable initializers was stripping
away GEPOperators when checking that the type of the initializer mathed the
expected type of a structure field, which led to many cases being marked as
"Unsafe pointer store", because these values were then tested as the type
of the global variable, instead of the field type the GEP produced. This
change adds logic to handle the GEPOperator directly.