CMPLRLLVM-30334: Fix for missing DTrans metadata and different array types (#958)
This change set adds a check when collecting information from
a DTrans type and the metadata is missing for it. If the pointers
are typed then it will return and will collect information from
the element type, else it will assert.

Also, it fixes the conditional that checks if two arrays/vectors
where dereferenced correctly.