[NFC] CMPLRLLVM-31003: Fix for memory leak in Intel partial inliner (#994)
This change set fixes a memory leak when running the
Intel Partial inliner. The
BranchProbabilityInfoWrapperPass and
DominatorTreeWrapperPass are function passes. Making these
passes as 'required' in a Module pass will compute the analyses
for all functions and the memory may not be freed correctly
at the end. This change set makes them not 'required'. Each
analysis will only be computed by the function actually that
needs it and is ready to be transformed.