[Driver] Only allow -mllvm -loopopt when -xARG is used (#460) (#940)
Improve filtering for passing -loopopt when -mllvm -loopopt is used
from the command line.  Do not allow it to be passed through unless
-xARG is also used.
Addresses:  CMPLRLLVM-31665
This is the second attempt at this (after a revert) as it was found
that the post-commit testing from the original xmain checkin showed
problems.  Re-starting the post-commit testing showed better results
demonstrating that these changes should be OK.