Add a while loop around the calls to __intel_cpu_features_init_x (#1030)
CMPLRLLVM-30527 points out that __intel_cpu_features_init_x doesn't work
well in threaded environments as it isn't re-entrant.  ICC likely has
the same problem, but it is aleviated by the fact that it wraps the init
call in a while-loop based on whether it is already initialized.

Presumably just an 'if' statement would be sufficient, but in order to
match ICC's behavior, this patch adds the loop around the call.