[CMPLRLLVM-32215] [DTrans] Provide determinism in resolving dominant types involving 3 or more type aliases (#1011)
When the DTrans type alias set contains 3 or more type aliases there is a
possibility that the computation of the dominant type will depend on the
order the elements are evaluated in. This occurs when one of the types is
a dominant type of the other two, but the other two do not dominant each
other. This is a rare condition that was observed with the new DTrans
analysis framework which resused code from the original DTrans analysis
framework, so is being fixed in both locations.