[X86] Align and insert JMP for short inner loops to help improve DSB hit. (#723)
If ChainBB is in a "outer" loop, and its fall through MBBs are a short
"inner" loop. It means that both "outer" and "inner" loops have high
probability to execute and ChainBB and its fall through MBB would have
DSB conflict.
Then adding a JMP between them and aligning them to make them allocate
in different DSB set, so that to avoid DSB conflict and improve DSB hit.