CMPLRLLVM-32280: Ensure that Threshold matches tested value if (#1068)
DecidedByCostThreshold is used.

The community removed code of the form:
if (ShouldInline.isSuccess() && CA.getCost() >= CA.getThreshold())
    return InlineCost::getAlways("empty function");
which would ensure that we would get a successful result even when the
CA.getCost() >= CA.getThreshold(). This caused an inconsistency when
the cost threshold test was performed as:
    Cost < std::max(1, Threshold)
but then later whether inlining had failed was tested using:
    explicit operator bool() const { return Cost < Threshold; }
Setting the threshold to the tested value lets us get a consistent result.