Ensure EmitFakeLoadForRetPtr removes from the alias scope struct. (#1166)
As reported in CMPLRLLVM-32326, when the FakeLoad mechanism is in play
and a fake load is emitted that causes the alias scope metadata to be
added, we end up having a use-after-free.  This shows up as a cast from
null issue, because the memory location is re-used for a future
instruction.

This patch fixes this issue by making sure that the Alias scope object
has its reference to the Instruction removed when we remove the
Instruction from the function.

The example itself otherwise does not seem to result in any of the alias
metadata being emitted, so the test is just to validate that it no
longer crashes.