CMPLRLLVM-30334: Update the IRMover to collect opaque pointers information from the llvm::Context (#1192)
This change set uses the llvm::Context to check if opaque pointers are
available. The check will be used to validate if incomplete metadata
is allowed to used, which is only available for typed pointers. Also,
this change set fixes the warning error print messages to only show
if the destination module is not empty, and adds a debug print for
tracing which source module produces metadata loss.