[NFCI][VPlanAlignmentAnalysis] Fix for CMPLRLLVM-31997 (#1152)
Assert happens for Align(0) because AlignmentAnalysis tries to propagate
alignment using the same base address. What happens is that the ptr difference
is 0, which results in known bits of 0. Then, when the alignment is propagated
using Align A(1ULL << KB.Zero.countTrailingOnes()), we end up with Align(0) and
an assert. Skip alignment propagation in these cases since there's no point in
doing it and there's no reason to record anything in CongruentMemrefs.