[OpaquePtr] Work around opaque pointers in SCCP solver. (#1199)
This usage was added for CMPLRLLVM-22930, and it appears that the changes made
there are an example of the sort of thing that opaque pointers will fix: there
would need to be an additional bitcast inserted somewhere whose omission causes
a compiler crash without opaque pointers. It's plausible that when opaque
pointer mode is irrevocably enabled, all of the code added for CMPLRLLVM-22930
could be ripped out.