CMPLRLLVM-32526: Ensure that inline report emitted when -qopt-report=3 is used does not seg fault when argument promotion is run before inlining. (#1233)
Whern argument promotion is run before inlining, the inlining report does
not have a module estabilshed. The module is only present in the inlining
report when the inliner is running. Since argument promotion can clone a
function, new call sites can be created and must be supplied with a module
so that it can be dumped when the inlining report is dumped at -qopt-report=3.
It is easier to get the module from the cloned callsites rather than pass
it down explicitly in each case that it is needed, so that is what I am doing
in this fix.

This is not an issue in the metadata based inlining report as it gets the
name of the module from the metadata.