[Driver] Add ability to enable all options for dpcpp (Linux based on Win) (#1235)
When using dpcpp on Windows we allow for both Linux and Windows style
options to be accepted so we can transition MSVC users to use dpcpp-cl
instead of dpcpp.  The initial step of this transition uses clang-cl
as the base, causing MSVC defaults to take hold (for instance needing to
enabling exception handling on the command line).  Add hooks to the driver
to allow for MSVC options when using clang and clang++, so we can have
Linux based defaults instead of MSVC.