[DTrans - OpaquePtr] Fixes for missing inferred types by PtrTypeAnalyzer (#1268)
When types need to be inferred based on users of the Value all users need to be
walked, so the PtrTypeAnalyzer would cache results for subsequent calls.
However, this caused a problem because in some cases the cached results were
incomplete because these intermediate results would depend on some other
value getting fully analyzed, which may depend on the initial value being
inferred. This updates it to avoid caching these results.
Also, avoid setting an inttoptr as 'unhandled' if it is reached by a
recursive analyzeValue call while the value is still being inferred.