CMPLRLLVM-32240: Part 1: Opaque pointer friendly padded malloc and padded pointer propagation. (#1250)
This is an initial change set that covers the LIT tests
paddedmalloc_safety*.ll.

I used Chris's adaptor to convert the code in DTransPaddedMalloc.cpp and
PaddedPtrPropagation.cpp. In PaddedPtrPropagation.cpp, I removed the checks
for padded malloc type propagation in getPaddingForValue(Value *V) and in
InFunctionPaddingResolver, because they are not needed and we are trying to
limit the use of the DTrans Metadata outside of DTrans.

In the DTransSafetyAnalyzer, I ported over the code to do the single alloc
function analysis, using Chris's new alloc analyzer.