[DTrans - OpaquePtr] Fix crash in DTransSafetyAnalyzer processing of indirect function calls (#1293)
The analysis for indirect function calls was crashing for vararg functions because
it was not checking the number for types available in the DTrans metadata about
the indirect function call before accessing it. Also, add a check that the DTrans
metadata information is consistent with the callsite when analyzing the indirect
call for matching functions.