[VPlan][NFC] Guard off some tests with SW_ADVANCED macro. (#1292)
As some CM code is guarded off in xmain-shared -like projects we treat related
LIT tests the same way. Update script has to be aware of SW_ADVANCED mark up
to keep it intact when updating CHECKs in LIT tests.
Also fix VF, to not depend on heuristics in xmain vs xmain-shared.