CMPLRLLVM-32240: Part 2: Opaque pointer friendly padded malloc and padded pointer propagation. (#1339)
This updates the LIT tests OpaquePtr/field-single-alloc-*.ll to run with both
the old and new pass managers and adds opaque pointer versions of these.
The opaque pointer version for field-single-alloc-func06.ll is omitted
because the test case indicates it is no longer relevant. There is also a
fix to DTransSafetyAnalyzer.cpp to handle the case where a single alloc
function is invalidated because a memset sets the value of the saved pointer
in a structure field to a non-zero value.