Replace the ExplicitEVEXPrefix with ExplicitPrefix. (#1348)
It seems we don't need ExplicitEVEXPrefix flag at all. So I replace this
flag to ExplicitPrefix which indicates:
No prefix is required for input, but the output is always with {vex} or
{evex} prefix. This flag should be added to most undisclosed instructions.

I will add this flag to MEMADVISE instructions only this time to reflect the
change. I will add this flag to all new isa in one new NFC patch.