[CMPLRLLVM-32674] [DTrans - OpaquePtr] Fix for assertion hit during DTrans safety analyzer analysis of global var initializer (#1373)
The PtrTypeAnalyzer was failing to construct ValueTypeInfo objects for some
functions leading to an assertion in the DTransSafetyAnalyzer when the value
was used as part of a global variable initializer. Fixing that exposed that
some function types, such as those used in the test
Transforms/paddedmalloc_safety04.ll, could not be represented. This updates
the DTransType system to include the LLVM 'token' type to allow the function
type to be represented.