CMPLRLLVM-32240: Part 4: Opaque pointer friendly padded malloc and padded pointer propagation (#1439)
This is a conversion to opaque pointers of a first set of LIT tests from the
PaddedPtrPropagation directory. I needed to generate the metadata for some of the
external functions in invoke.ll becuase the C/C++ front end was not doing that. I
will take this up with Erich Keene when he gets back. (Thanks for the pointers, Chris.)

I updated the .rst file for opaque pointers to include an explanation of the
"intel_dtrans_func_index"es.

I have also fixed a bug where I forgot to test in the legacy pass manager for the use of
the appropriate type of DTrans analysis.