[CMPLRLLVM-32827] [DTrans - OpaquePtr] Fix for non-deterministic behavior on getDominantType function (#1494)
The pointer type analyzer function getDominantType() incorrectly allowed a generic
pointer type to override the non-generic pointer type when searching for a dominant
type. Also, update the pointer type analyzer trace that reported the dominant type
to report for all cases, instead of just when there is a pointer to an aggregate
type.