[VPlan][SVA] Introduce Scalar and Vector specifications of SVA. (#1438)
The class that represents SVA for VPlan is split so SVA has
different classes for Scalar SVA and Vector SVA. Scalar SVA
applies for VF = 1 case only and doesn't do any real analysis
but always returns true for 'FirstScalar' enquires and false
otherwise. It is done in preparation to invoke CM for VF = 1
very early in pipeline, when many analyses required for SVA
are not available.