[VPlan][HIR] Fix bug in setting index operand of wide VLS memref. (#1523)
The index operand on the base address of a wide VLS memref should be
scalar in-order to emit a wide consecutive load/store. Changes in
PR #1283 incorrectly changed the type of index from scalar to vector
leading to incorrect broadcast + gather/scatter pattern. This patch
addresses this bug and also renames associated parameter in
createMemrefFromBlob interface to avoid future confusions.

JIRA: CMPLRLLVM-32755