[xmain-web] Add XFAIL back to thinlto-distributed-newpm.ll (CMPLRLLVM-32876) (#1557)
This test has been marked as XFAIL in the past because of CMPLRLLVM-344, but
XFAIL was recently removed from this test during pulldown conflict resolution
(please see da9d1e4be84bd05149e06359782ad7aa0ef3e536) because it has started
passing. Unfortunately that did not last for a long and it has started failing
again in couple days after removing XFAIL. Adding it back to resolve this fail.