[NFC] Clone CallBase::getFnAttr to CallBase::getCallSiteOrFuncAttr (#1618)
INTEL_CUSTOMIZATION was implemented in time when community didn't have getFnAttr
but once they implemented it they gave it different semantics. The plan is to
have our customized version named differently but it will be done in two steps
because the changes crosses multiple repositories. This first patch creates a
new name and will allow other repos to be updated without breaking things.