CMPLRLLVM-32596: Identify downcasting with opaque pointers (#1631)
This change set enables the support for identifying downcasting
when there are opaque pointers in the IR. The basic idea is that
we are going to collect the dominant type using the DTrans
metadata and the DTrans pointers analyzer. If we can find the
dominant type then we need to prove that the list of possible
destination types are the same as the dominant type. If we find
a derived type in the list, then we have a possible downcasting
and the assume call needs to be removed.

If the dominant type can't be collected then we add the default
case as part of the multiversioning when devirtualizing a
virtual call.