[NFC][VPlan] Fix for CMPLRLLVM-32950 (#1637)
Prevent unsupported vconflict loops from leaking out to LLVM-IR.
Currently, histogram idiom is the only vconflict idiom supported in
VPlan and this happens only on the HIR side. Other idioms are rejected
by VPlanDriverHIR, but the loop directives were not removed. Then,
VPlanDriver picks these regions up and vectorizes them, resulting
in runfail. This patch removes the directives when the idiom is not
supported.