[NFC][VPlan][AllZeroBypass] Fix for CMPLRLLVM-33082 (#1702)
RegionBlocks was not initialized for each loop during region collection.
Somehow this assertion was not triggered previously, but from a stability
perspective, reusing RegionBlocks for loops would be unlikely to have
caused any actual stability issues because it is used for only two
purposes: 1) verification that the loop region is correct 2) checking to
see if the loop region is already included in another region with the same
predicate. In the case of #2, this would just mean we have an extra bypass.