[DTRANS] CMPLRLLVM-32994: Fix for segmentation fault when analyzing pointers to opaque structures (#1698)
This change set fixes a possible segmentation fault that could
be produced by the DTrans pointer analyzer. The pointer
analyzer assumes that all the structures will have at least one
field, when there could be cases when the structure is opaque.
If the structure doesn't have a body, then it will produce a
segfault when collecting a field.