[SYCL] Generate and install stripped PDBs for SYCL libraries (#1318)
Adds stripped PDB files for SYCL library and the PI plugins when
building with MSVC. Full PDB files will also be generated, but only the
stripped variants will be installed.

The stripped PDB files will only be generated and installed if the used
linker supports the /PDBSTRIPPED options. LLD does not currently support
this option. If the stripped PDB is not generated, no PDB files are
installed for the SYCL libraries and PI plugins.

JIRA: CMPLRLLVM-28931