CMPLRLLVM-33159: Fix for IRMover to support structures without mangled names (#1709)
There is a chance that a transformation in the backend adds
a new structure without mangled named. The issue is that
the IRMover can skip this structure when merging the types
with DTrans metadata. This change set will add the support
for checking if the structure names can be merged by checking
the names without extra numbering when there is no mangled
names information.