[DTRANS] Emit DTRans info for the 'runtime functions' (#1724)
As reported in cmplrllvm-32785, all of the exception-related runtime
function calls weren't being annotated correctly since they are
generated differently. This results in the optimizations having to be
pessimistic. Looking through this found that this problem exists for
dyn_cast and static-init as well.

This patch correctly sets the metadata for these runtime functions.

Note that the MSVC 'throw' required a bit of extra work, since its
signature includes a special 'runtime type' that ALSO needed to be
annotated.