[OpenMP] Ensure static/inline declare target with target regions are compiled on device (#1712)
The logic that existed treated functions marked 'declare target'
normally and they would be emitted when used. Since there may be
no use they would not be emitted if inline (for example).  But
when there is no 'declare target' the function would be emitted
since it is forced in that case.

This change adds these 'declare target' function with target
regions to the deferred list and they can be emitted normally.

Fixes: CMPLRLLVM-26882