[Dynamic LoopOpt] Implement support in Old Pass Manager (#944)
The implementation will be different (and much simpler) in the New PM because
the Function/Loop PM is itself a pass there, so it's easy to wrap that with a
thin helper class that will early return and not apply the whole set of passes
at once.

In the Old PM the PassManagerBase class hierarchy doesn't expose that,
unfortunatelly. Only deep in the implementation of the PM the actual decscendent
of a Pass is created based on the object of the PassManagerBase and it's too
late and inconvenient for us to exploit that. Instead, we customize the Pass
itself to assign a "Limiter" to it and also add aditional checks inside
Function/Loop pass managers implementation to not run the pass if it's limited
and function's attributes indicate that the pass needs to be skipped on that
function.

Note that in order to do that we need to carry the enabled loopopt pipeline on a
given function via dedicated attribute, i.e. not as part of TTI-like analysis as
that wouldn't be accessible from that place.

Co-authored-by: Andrei Elovikov <andrei.elovikov@intel.com>