[DTrans - OpaquePtr] Cleanup resolving DTransType on intrinsic and library functions in the PtrTypeAnalyzer (#1747)
The initial development of the PtrTypeAnalyzer used some hardcoded types for
intrinsic and library functions before the work for the front-end to produce
metadata was done. This cleans up the implementation to isolate this information
into a new class DTransLibraryInfo. Because metadata is present for most
funtions now, this class only needs to expose type information for functions
that may be inserted by transformation passes, so removes much of the
previous hardcoding that was contained within the PtrTypeAnalyzer.

This changeset only provides a subset of the __kmpc functions
produced by PAROPT which are present in 638 and 657. Additional functions
produced for other cases can be added as needed.