review/grokos/topic (#1782)
* [SLP][NFC]Update test checks, NFC.

* [SYCL] Conditionally enable diagnosing that requires integration header (#5044)

https://github.com/intel/llvm/pull/4945 introduced a new diagnostic that
works correctly only if integration header is included. However it is
possible that `-fsycl-is-host` flag can be used without integration
header, so the diagnostic should be issued only if integration header is
included. This patch adds new cc1 option that helps to understand that.

* Align driver tests with changes from upstream

The new option is added by the driver if integration header is included.

* [ScalarEvolution] Fix crash where Map Entry was removed due to earlier invalidation (#1780)

Co-authored-by: Alexey Bataev <alexey.bataev@intel.com>
Co-authored-by: Mariya Podchishchaeva <mariya.podchishchaeva@intel.com>
Co-authored-by: Chenyang Liu <chenyang.liu@intel.com>