Fix incorrect read/write to local values in non-kernel functions (#116)
The local memory will be rebased when a non-kernel function is called.
But the value of locals are neither copied to the local memory allocated
for the callee nor written back.

This patch stores the address of locals to local memory of callee, and
thus have to load the address before load the value in callee. In this
way, load(s)/store(s) will operate on original memory location.