CMPLRLLVM-32875: Look in "bin" for SYCL programs (#1781)
The current Driver framework adds the driver's installed directory as a
program path to look for tools in, but the driver is discovered to be
clang rather than icx. This means that we look for tools in bin-llvm
rather than bin. This patch adds "../bin" as another path to search for
tools.