CMPLRLLVM-33272: Fixes due to switching order of normal and aggressive (#1815)
InstCombine.

Changes to PassBuilderPipelines.cpp fix an incorrect conflict merge.
The number of times normal InstCombine was invoked was increased by 1.
The correct change should have been to simply interchange the order of
normal and aggressive InstCombine.

The change to pr52289.ll ensures that the new pass manager is invoked, even
when it is not the default in xmain.

There was also a failure with the new community LIT test
  Transforms/PhaseOrdering/X86/merge-functions.ll
This is due to Intel customizations in:
   llvm/lib/Transforms/Utils/SimplifyCFG.cpp
made by Farhana Aleen on 23 Apr 2015. I suggest opening a new JIRA item
for this and assigning it to LLVM Scalar Optimizer. See the CMPLRLLVM-33272
for details.