[NFC] [DTrans] Make FieldInfo objects non-copyable (#1813)
FieldInfo objects are members of the StructInfo class, and functions
from StructInfo that retrieve them always return them by reference. It
does not make sense to allow them to be copied because this could lead
to code updating a copy instead of the element that StructInfo uses.
Fix locations in DTransInfoAdaptor where FieldInfo was returned as a
copy.