[VPlan][CM] Drop out of loop blocks costs. (#1803)
We don't have to scale costs of the blocks that do not execute inside of
the loop up by TC factor. Solid solution for that might imply infrastruture
changes to support block execution counts and solution for cost overflow
problem (when block cost is multiplied by loop TC). The temporal solution
is to drop the costs of the blocks that do not belong to the loop.