[DTrans] Update -dtrans-print-types to print 'End' message after TypeInfo records (#1880)
This improves testability of the DTrans safety data messages by making sure the
'Safety data' corresponds to a specific type. Before this change it was possible
that the last 'Safety data' message check line could match against a different
type, if there were additional structures or array types reported that did not
have explicit checks. A side-effect of the change
https://github.com/intel-restricted/applications.compilers.llvm-project/pull/1191
was that all types get populated into the TypeInfo collection, which includes
some builtin TypeInfo types used by the PtrTypeAnalyzer, resulting in additional
'Safety data' lines being reported. This change also gives some additional
flexibility for testing because all types do not need to have CHECK lines for
cases where those types are not of interest for the functionality under test.

This change updates the existing tests to use this 'End' delimiter.