CMPLRLLVM-32953: Fix for compile time when using DTrans metadata and merging IR (Part 1) (#1893)
This change set fixes the compile time overhead introduced by the
IR mover with DTrans metadata. The issue is that the IR mover
was loading the DTrans metadata from the destination module,
as well collecting the module structures, on every iteration
that loads a source module. The solution is that there is going
to be a DTrans type manager for the destination module and the
destination DTrans types will be updated as needed.