[DTrans - OpaquePtr] DTransSafetyAnalyzer to use info from DTransLibraryInfo class (#1940)
The change in https://github.com/intel-restricted/applications.compilers.llvm-project/pull/1747 allowed some library functions to not have DTrans metadata.
However, when the DTransSafetyAnalyzer analyzed calls to these functions,
it was still producing "Unhandled use". This updates the DTransSafetyAnalyzer
to avoid redoing the behavior that the PtrTypeAnalyzer does in VisitModule to
determine a DTransFunctionType for a Function, and just retrieves the
information from the PtrTypeTypeAnalyzer.