[CMPLRLLVM-27696] [DTrans] Fix DeleteField pass to update size arguments for calls involving pointers to arrays of structures (#1944)
When there is an explicit array of structure elements that are being
transformed by the DeleteField pass, the size argument used in the calls
that are of interest needs to be updated. Previously, the size was only
updated when there was an implicit array detected based on an allocation
was a multiple of the structure size, and the test was marked as XFAIL.
Also, the test was not ported to work with the new opaque pointer
framework pass.