[DTrans - OpaquePtr] PtrTypeAnalyzer to try to infer argument types for functions without metadata (#1973)
This is related to CMPLRLLVM-32677 for DTrans being able to analyze IR following
PAROPT transformations that split part of a function out to be used via the
openMP runtime. In these cases, there is no metadata on the new function, but
DTrans can infer the argument types based on the usage.