CMPLRLLVM-33510: Check if type supports LLVM::Constant before folding (#1975)
The LLVM community added a constant folding process in the global
optimizer (https://reviews.llvm.org/D114889). The problem is that
it doesn't check if the Type of the Value supports LLVM::Constant
and it can produce an assertion.

This change set adds a check before the constant folding to make
sure that the type supports LLVM::Constant before doing the
transformation.