[NFC][VPlan][LLVM IR] Add cl::opt to use original BBs' names (#1977)
This updates vplan_predicator.ll test as it benefits from the change the most.

Unfortunately, I wasn't able to drop the VBBB capturing in
intel_update_vplan_checks.py because we uniqify names across whole compilation
and not for a given VPlan. That means that names differ if a function is
compiled standalone or as a part of a bigger module.