CMPLRLLVM-33585: Fix for passing whole program utils through the new pass manager and update whole program read test cases (#2016)
This change set fixes the issue where the information of the
whole program utils wasn't passed to the whole program analysis
in the new pass manager. The issue is that the WPUtils was set
after creating the whole program analysis pass.

Also, this change set updates the test cases for whole program
read in GOLD and LLD linkers. The test cases are ELF only.