[VPlan][CM] Relax certain intsructions pattern match in VPlan CM. (#2140)
The change removes checks for red-init and red-fini VPIstructions when
searching for saturated absolute difference pattern. This step is required
for moving VF=1 cost moddeling earlier in pipeline, even before reduction
entities are detected and added to VPlan.

There is alternative approach, which is to make detection to happen for VF > 1
only Plans (not VF = 1 Plan as it currently is) and add some cost to them rather
than reducing the cost of VF = 1 Plan (as it currently does).